Refactor components and rename route path
